refactor(runtime): extract recovery guard orchestration tranche 5 (#3129)
